requirements.txt
langchain==0.3.0
langgraph==0.2.0
langchain-openai==0.2.0
chromadb==0.5.0
streamlit==1.35.0
sqlalchemy==2.0.0
pandas==2.2.0
python-dotenv==1.0.0
